+++ changelog.md
@@ -15,6 +15,8 @@
 
 *	2021.4.9
 
+	+	organizeTokens.ts: refined system splitting by direction decorator tokens.
+
 	+	organizeTokens.ts: split staff tokens by source direction decorator.